# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/users/{user_id}/workout-stats:
    get:
      tags: [Analytics]
      summary: Get user workout stats
      operationId: getUserWorkoutStats
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Workout statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutStats"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/activity-calendar:
    get:
      tags: [Adherence]
      summary: Get user activity calendar
      operationId: getActivityCalendar
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: start
          schema:
            type: string
            format: date
        - in: query
          name: end
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Map of dates to activity types.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/streaks:
    get:
      tags: [Adherence]
      summary: Get user streaks and adherence summary
      operationId: getUserStreaks
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Reference date for streak calculation (defaults to today).
      responses:
        "200":
          description: Streaks and adherence summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreaksResponse"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /v1/exercises/{id}/history:
    get:
      tags: [Exercises]
      summary: Get exercise history
      operationId: getExerciseHistory
      parameters:
        - $ref: "#/components/parameters/ID"
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Ordered past sessions for the exercise.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedExerciseHistory"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/workout-templates/{id}/apply:
    post:
      tags: [Workout Templates]
      summary: Apply a template to create a new workout
      operationId: applyTemplate
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyTemplateRequest"
      responses:
        "201":
          description: Workout created from template.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    UserID:
      in: path
      name: user_id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [title, status, error]
      properties:
        type:
          type: string
          description: Problem type URI. Validation errors use a dedicated type; generic errors use `about:blank`.
          example: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error category.
          example: Bad Request
        status:
          type: integer
          description: HTTP status code for the error response.
          example: 400
        detail:
          type: string
          description: Human-readable explanation of the specific error instance.
          example: invalid email or password
        error:
          type: string
          description: Backward-compatible alias of `detail`.
          example: invalid email or password
        errors:
          type: object
          description: Field-specific validation messages when the request fails validation.
          additionalProperties:
            type: string
          example:
            email: email is required
            password: password must be at least 8 characters
...
    ApplyTemplateRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
...
    WorkoutStats:
      type: object
      required: [total_workouts, total_duration, total_volume, total_sets, avg_workout_duration, workout_types]
      properties:
        total_workouts:
          type: integer
        total_duration:
          type: integer
        total_volume:
          type: number
        total_sets:
          type: integer
        avg_workout_duration:
          type: number
        workout_types:
          type: object
          additionalProperties:
            type: integer
    StreaksResponse:
      type: object
      required: [streaks, adherence_summary]
      properties:
        streaks:
          $ref: "#/components/schemas/Streaks"
        adherence_summary:
          $ref: "#/components/schemas/AdherenceSummary"
    Streaks:
      type: object
      required: [workout_streak, meal_streak, weigh_in_streak]
      properties:
        workout_streak:
          type: integer
        meal_streak:
          type: integer
        weigh_in_streak:
          type: integer
    AdherenceSummary:
      type: object
      required: [days_7, days_30, days_90]
      properties:
        days_7:
          type: number
        days_30:
          type: number
        days_90:
          type: number
    ExerciseHistoryEntry:
      type: object
      required: [date, workout_id, workout_type, sets, volume, estimated_1rm]
      properties:
        date:
          type: string
          format: date-time
        workout_id:
          type: string
          format: uuid
        workout_type:
          type: string
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSetData"
        volume:
          type: number
        estimated_1rm:
          type: number
    WorkoutSetData:
      type: object
      required: [weight, reps, rpe, completed]
      properties:
        weight:
          type: number
        reps:
          type: integer
        rpe:
          type: number
        completed:
          type: boolean
...
    UserSummary:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
...
    ExerciseSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    WorkoutSet:
      type: object
      required:
        - id
        - workout_exercise_id
        - set_number
        - reps
        - weight
        - rpe
        - rest_seconds
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workout_exercise_id:
          type: string
          format: uuid
        set_number:
          type: integer
        reps:
          type: integer
        weight:
          type: number
          format: double
        rpe:
          type: number
          format: double
        rest_seconds:
          type: integer
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkoutExercise:
      type: object
      required:
        - id
        - workout_id
        - exercise_id
        - order
        - sets
        - reps
        - weight
        - rest_time
        - notes
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workout_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
        order:
          type: integer
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: number
          format: double
        rest_time:
          type: integer
        notes:
          type: string
        exercise:
          $ref: "#/components/schemas/ExerciseSummary"
        workout_sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Workout:
      type: object
      required:
        - id
        - user_id
        - date
        - duration
        - notes
        - type
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        duration:
          type: integer
        notes:
          type: string
        type:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"
        workout_exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginationMetadata:
      type: object
      required: [page, limit, total_count, total_pages, has_next]
      properties:
        page:
          type: integer
          description: Current page number.
        limit:
          type: integer
          description: Maximum number of records requested per page.
        total_count:
          type: integer
          description: Total number of records matching the query.
        total_pages:
          type: integer
          description: Total number of pages available for the current limit.
        has_next:
          type: boolean
          description: Whether another page of results exists after the current page.
...
    PaginatedExerciseHistory:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseHistoryEntry"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT